packages/snap/src/core/polyfills/intl/Collator.test.ts
import { Collator } from './Collator';

describe('Collator', () => {
  describe('constructor', () => {
    it('uses default locale when none provided', () => {
      const collator = new Collator();
      expect(collator.resolvedOptions().locale).toBe('en');
    });

    it('uses first locale from array', () => {
      const collator = new Collator(['fr', 'en']);
      expect(collator.resolvedOptions().locale).toBe('fr');
    });

    it('uses default options when none provided', () => {
      const collator = new Collator();
      expect(collator.resolvedOptions()).toStrictEqual({
        locale: 'en',
        usage: 'sort',
        sensitivity: 'variant',
        ignorePunctuation: false,
        numeric: false,
        caseFirst: 'false',
      });
    });
  });

  describe('compare', () => {
    it('compares strings basic', () => {
      const collator = new Collator();
      expect(collator.compare('a', 'b')).toBeLessThan(0);
      expect(collator.compare('b', 'a')).toBeGreaterThan(0);
      expect(collator.compare('a', 'a')).toBe(0);
    });

    it('handles ignorePunctuation option', () => {
      const collator = new Collator('en', { ignorePunctuation: true });
      expect(collator.compare('hello!', 'hello')).toBe(0);
      expect(collator.compare('hi.there', 'hithere')).toBe(0);
    });

    it('respects numeric sorting', () => {
      const collator = new Collator('en', { numeric: true });
      expect(collator.compare('2', '10')).toBeLessThan(0);
    });
  });

  describe('supportedLocalesOf', () => {
    it('handles single locale', () => {
      expect(Collator.supportedLocalesOf('en')).toStrictEqual(['en']);
    });

    it('handles array of locales', () => {
      expect(Collator.supportedLocalesOf(['en', 'fr'])).toStrictEqual([
        'en',
        'fr',
      ]);
    });
  });

  describe('resolvedOptions', () => {
    it('returns merged options', () => {
      const collator = new Collator('fr', {
        usage: 'search',
        sensitivity: 'base',
        ignorePunctuation: true,
        numeric: true,
        caseFirst: 'upper',
      });

      expect(collator.resolvedOptions()).toStrictEqual({
        locale: 'fr',
        usage: 'search',
        sensitivity: 'base',
        ignorePunctuation: true,
        numeric: true,
        caseFirst: 'upper',
      });
    });
  });
});
